Added load factor change into builder functionality

+++ src/main/java/org/apache/datasketches/filters/quotientfilter/QuotientFilter.java
@@ -18,6 +18,7 @@
  */
 
 package org.apache.datasketches.filters.quotientfilter;
+import static org.apache.datasketches.filters.quotientfilter.QuotientFilter.DEFAULT_LOAD_FACTOR;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -31,11 +32,11 @@ import org.apache.datasketches.filters.common.HeapBitArray;
 
 public class QuotientFilter extends Filter {
 
-  public static final float DEFAULT_LOAD_FACTOR = 0.8f;
+  public static final double DEFAULT_LOAD_FACTOR = 0.8;
 
   int lgQ_;
   int numFingerprintBits_;
-  float loadFactor_;
+  double loadFactor_;
   int numEntries_;
   int numExpansions_;
   BitArray bitArray_;
@@ -50,13 +51,12 @@ public class QuotientFilter extends Filter {
     this(lgQ, numFingerprintBits, DEFAULT_LOAD_FACTOR);
   }
 
-  public QuotientFilter(final int lgQ, final int numFingerprintBits, final float loadFactor) {
+  public QuotientFilter(final int lgQ, final int numFingerprintBits, final double loadFactor) {
     lgQ_ = lgQ;
     numFingerprintBits_ = numFingerprintBits;
     loadFactor_ = loadFactor;
     bitArray_ = makeFilter(getNumSlots(), getNumBitsPerEntry());
     numExpansions_ = 0;
-    //hash_type = XxHash.hashLong ; //HashType.xxh;
   }
 
   public boolean rejuvenate(final long key) {
